backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
SentenceTransformer
